requirements.txt
tensorflow
flask
pillow
virtualenv